Implement missing SD plugin variables.
Seems when designing the SD plugin variables a copy was made of the
stuff available in the director. That of course doesn't work to well in
the SD as some info is just not available there. Implemented now as many
vars which can be filled with valuable info and removed the others
which make little sense.

Also did a cleanup of all plugin files to output the debug messages
generated for a variable retrieval to look the same.