do not install files as conffiles in /etc/bareos
If we install those files as conffiles, we are not allowed to replace
passwords etc in them.

This change installs the files into /usr/share/bareos-common by default
and uses them templates for the generated files in /etc/bareos in postinst.

Ideally, we would register them with ucf and merge the newer versions in
the future, but this task is left to future-me or someoneelse who has the
needed time.

Closes: #767819

Signed-off-by: Marco van Wieringen <marco.van.wieringen@bareos.com>