Tweak queries a bit.
Create a second query for finding the last full backup when no specific
pool is specified as then we don't need to JOIN the JobMedia and Media
table into the query as we don't select on the Media's PoolId.

Also drop the JOIN on Client as we don't select any column from that
table anyway and the query is also kind of clumbsy by specifying twice
the ClientId instead of using a single WHERE on the ClientId and use
an AND Client.ClientId = Job.ClientId.

Fixes #363: zero-byte Full backups are ignored for restore