Fix FD plugin handling.
Seems we broke the FD plugin handling while converting it to the new
plugin framework. These changes make the regression script at least
working again and also cleanup the test if we are processing an entry
of a plugin by a new method in the JCR is_plugin() which checks for
the cmd_plugin or option_plugin boolean.